Make Persister generic to avoid casting

+++ gemma-core/src/main/java/ubic/gemma/core/datastructure/matrix/MatrixWriter.java
@@ -163,7 +163,7 @@ public class MatrixWriter {
                     //noinspection RedundantStringOperation // being explicit
                     buf.append( "" );
                 } else if ( val instanceof Double ) {
-                    buf.append( String.format( "%.3g", ( Double ) val ) );
+                    buf.append( String.format( "%.3g", val ) );
                 } else {
                     buf.append( val );
                 }
